Capitalize GitHub everywhere it makes sense to do so
Officially the name of the site is "GitHub", not "Github". In filenames,
URLs, and so on, it makes sense to have this be lowercase, but the name
of the site shouldn't be output to users without the capital H.

Also changed the internal name of the config section, but that won't
affect existing users.