readme: update line numbers for webhook permission check
In the future, the readme probably should use an alternate means of
indicating why these are necessary, because this link is very fragile.
Either we start referencing a specific code revision (thus potentially
linking to very outdated code, over time), or we stop linking from
documentation into the code like this entirely.