webhook: Process payload in a separate thread
GitHub seems to log timeouts for webhooks it sends to this plugin pretty
often—or at least, it does on the instance we use in the `#sopel` chan
at freenode. The timeout on GitHub webhooks is pretty short (10 seconds
tops, maybe shorter), so without actually profiling the code, I can
guess that processing payloads just takes too long sometimes what with
calling out to the URL shortener and all.

Doing the processing in a separate thread is an obvious trick to try. It
lets bottle send an HTTP response back to GitHub pretty quickly (only
needing to wait for a single database query in `get_targets()`).