main: update regexes to match GitHub URL bits
The username and repo name regexes weren't exactly accurate. They
allowed underscore (`_`) which GitHub will not, and did not accept dot
(`.`) in repo names even though GitHub allows that. (This was discovered
by posting a link to Sopel's website, whose repo is named `sopel.chat`,
and noting that the GitHub module didn't handle it as expected.)

With a bit of Google searching and some tweaks, we should have much more
robust URL matching now.