Update for Sopel 7 (requirements; URL handling)
We'll release v0.3.0 for Sopel 7.0+ only, so we can make use of the
fixes to URL handling (decorator) and eliminate manual registration of
callbacks that was required before to avoid conflicts with Sopel's
built-in title fetching plugin.

Using the actual URL decorator should also more-or-less fix #7 and the
issue-less bug we had when using `rule` where only one GitHub link of
any type per IRC message would ever be handled.