github: change API auth from query params to HTTP Basic
GitHub deprecated passing authorization via query parameters:
https://developer.github.com/changes/2019-11-05-deprecated-passwords-and-authorizations-api/

HTTP Basic Auth is the way forward, and they ignored (so far; it's been
a week) my support request asking if there would be a better way to
accomplish this that didn't involve OAuth or personal tokens. If this
too gets deprecated later, I'll be quite annoyed, as we don't WANT to
authenticate as a specific user in this plugin.