formatting: skip label/unlabel events without label in payload
As the comments say, if a label is deleted, GitHub will send 'unlabeled'
events for every issue/PR that had the deleted label, but the 'label'
object won't be included. Rather than emit a placeholder ('[unknown
label]' or something like that), skip these events. They can be pretty
spammy anyway, if a label used on many issues/PRs is deleted.