github: fix trigger groups in `.gh` command
I have no idea why this was using a fallback to group 1 (the command
name) instead of just detecting an empty/missing group 2 (or 3).

Now instead of a generic "Not found" error for missing argument, caused
by looking for a repo at `<nick>/<command_name>`, we have a proper usage
reply and an example decorator.