webhook: fix error on reconnect
The bot runs plugin `shutdown`s on disconnect from the IRC server. While
this method stops serving requests it doesn't actually close the socket;
the restarted webserver cannot then bind to the same host/port in order
to listen+serve requests.  This results in a `OSError: [Errno 98]
Address already in use`. To avoid this, the server needs to be closed
after `shutdown`.

As of this commit, the fix is valid for all Python>=2.7,>=3.5.