github: don't try to access issue state in comment payloads
Adding display of the issue state ('open', 'closed', or 'merged') and
type ('issue' or 'PR') broke displaying comments, because the API does
not return information about the issue itself when fetching a comment.
This is why the title is only appended `if ('title' in data)`. Had to
do basically the same thing for the 'state' field.