github: support inline references to issues/PRs in any repo
This accepts `user/reponame#123` in all cases.

As before, if the current channel has a default repo (set using the
`.gh-repo` command), `#123` will be treated as a reference to an issue
or PR in the configured default repo. Additionally, `reponame#123` will
now fill in the missing "username" portion from the configured default.

Implementing this feature required switching to named capture groups in
all URL patterns, because the `user` and `reponame` portions of the
inline references are optional. It seemed like a maintainability win,
anyway, so even if I just missed thinking of a way to do it with
numbered capture groups, I'm glad to have made the change anyway.