github: remove obsolete `secret` setting
Long replaced by `client_secret`, along with slightly janky (because it
auto-rewrites Sopel's config file, erasing comments) auto-migration code
to move the old `secret` value to the new place where it belongs.

Was meant to be removed in 0.3.0, but we went through the whole 0.4.x
series with this still present.