github: simplify getting PR merge status
`merged_at` is `null` if the PR is either open, or closed unmerged.

Not 100% certain that field is guaranteed to be populated for *every* PR
on GitHub, regardless of age, but I think it's acceptable to get rid of
the second API request in exchange for a small risk of giving slightly
incorrect output in a very small number of cases.